Fix all type errors
